# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  main.adminCreateUserRequest:
    properties:
      email:
        type: string
      locale:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
...
  main.adminUserView:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastConnectedAt:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
...
  main.roleRequest:
    properties:
      role:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: CONXTOR project
    url: https://github.com/conxtor/groupmaster
  description: |-
    REST API for ingesting, analyzing, and presenting selected WhatsApp and Telegram group conversations.
    The public deployment serves this API below the /api reverse-proxy prefix.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/conxtor/groupmaster
  title: CONXTOR Messaging Group Intelligence API
  version: 0.1.0
paths:
...
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.adminUserView'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List users
      tags:
      - administration
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.adminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create user
      tags:
      - administration
  /admin/users/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role and status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.roleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update user role/status
      tags:
      - administration
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: Authenticated CONXTOR session cookie.
    in: cookie
    name: wagi_session
    type: apiKey
swagger: "2.0"